Database class refactory (separated in two classes)
