+++ Sharp.Data/Sharp.Data.csproj
@@ -80,6 +80,7 @@
     <Compile Include="Databases\PostgreSql\PostgreSqlProvider.cs" />
     <Compile Include="Databases\PostgreSql\SqlColumnStructure.cs" />
     <Compile Include="DbFactory.cs" />
+    <Compile Include="DefaultDatabase.cs" />
     <Compile Include="ExceptionHelper.cs" />
     <Compile Include="Exceptions\DatabaseException.cs" />
     <Compile Include="Databases\DataProvider.cs" />
